+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.2.1] - 2022-05-01
 
--
+### Added
+
+- Dynamic worker choice strategy change at runtime.
 
 ## [2.2.0] - 2022-05-01